Remove duplication of constants
Before this we had the same constant names, with the same value across multiple components.  Now constants which are used in two places live in /components/constants.  Components that consume these import these constants directly, and then any constants that are specific to that component via a constants file in the root of that components folder.