Remove duplication of SMALL_WIDTH & SMALL_SCREEN consts
Previously these two consts had the same value, different names, but were used in very similar ways.  As such, it looks like these were duplicates.  In this commit we remove that duplication by having just one const (SMALL_WIDTH) in the global consts, and using that.  I chose the SMALL_WIDTH name because it looks like it's the width of the chart, not the width of the screen