Upgrade the dev dependencies related to sewing-kit
This upgrade includes all functionality served by sewing-kit, including
but not limited to, lint, format, and test. As part of this upgrade, we
upgraded the version of TypeScript and the version of Node (staying
within the 10.x family).

Finally, a few changes were made to be more like Polaris, including a
more optimized tsconfig.json for normal dev stuff and a
tsconfig.build.json for build stuff. This resulted in adding
"import type" and "export type" to places where we're dealing with
types.

As part of this change, I've run through sewing-kit's lint, format,
type-check, test, optimize, and playground functionality and all of them
appear to work as intended.