Add tslib to our dependencies
We actually depend on this package since the code that TypeScript
generates will transform some syntax into function calls. For example,
Array Spreading syntax gets converted to __spreadArray from tslib.

When we tried to use version 0.17.0 of this package in the web project,
we found that the version of tslib that gets resolved is actually 1.x
instead of 2.x which we depend on.

This change requests tslib version ^2.1 since that is when __spreadArray
was introduced: https://github.com/microsoft/tslib/releases/tag/2.1.0