Upgrade sewing-kit to 4.0.0
Replace our custom browserslist configuration with the standard Shopify
configuration to match what we do in other projects.

Also, updates the browserslist resolution to ^4.17 since this is what
sewing-kit needs and we can't remove the resolution (yet) because other
packages still depend on older versions, which are vulnerable to
security issue(s).