Change polaris-viz.shopify.io -> .com
`polaris-viz.shopify.io` does not seem to exist.
`polaris-viz.shopify.com` is interspersed throughout the code. Change
all references from `.io` to `.com` in the READMEs, etc

Did not change the `CHANGELOG` as it's historical

Finally, change the storybook definition in `RENDER_TOOLTIP_DESCRIPTION`
because it's used in the site to link back, and that link was broken.