fix: Expose Uplift theme
We need to be able to have two variations of the Uplift theme: a
standard and a "ghost" version.

As such we can't simply overwrite properties of the theme as shown
[here](https://polaris-viz.shopify.com/iframe.html?viewMode=docs&id=shared-themes-customizing--page#titleAnchor-81).

Instead we need to define multiple themes as shown
[here](https://polaris-viz.shopify.com/iframe.html?viewMode=docs&id=shared-themes-customizing--page#titleAnchor-86).

To do this with Uplift as a base however, the theme needs to be exposed,
which it currently isn't.

In theory this can be retrieved via the `@shopify/polaris-viz-core`, but
the same is true for the other themes and that would mean that
dependency becomes a direct dependency of the project, rather than
through `@shopify/polaris-viz`.

Instead, it's proposed here that we just expose the Uplift theme in a
manner similar to the others.